+++ webscout/Provider/__init__.py
@@ -37,6 +37,7 @@ from webscout.Provider.meta import Meta
 from webscout.Provider.Netwrck import Netwrck
 from webscout.Provider.Nvidia import Nvidia
 from webscout.Provider.oivscode import oivscode
+from webscout.Provider.OpenRouter import OpenRouter
 from webscout.Provider.PI import PiAI
 from webscout.Provider.QwenLM import QwenLM
 from webscout.Provider.Sambanova import Sambanova
@@ -53,11 +54,12 @@ from webscout.Provider.WiseCat import WiseCat
 from webscout.Provider.WrDoChat import WrDoChat
 from webscout.Provider.x0gpt import X0GPT
 
-from .Openai import OPENAI
+from .Openai import OpenAI
 
 # List of all exported names
 __all__ = [
-    "OPENAI",
+    "OpenAI",
+    "TypliAI",
     "AI4Chat",
     "AkashGPT",
     "Algion",
@@ -94,6 +96,7 @@ __all__ = [
     "Meta",
     "Netwrck",
     "Nvidia",
+    "OpenRouter",
     "oivscode",
     "PiAI",
     "QwenLM",